requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
python-dotenv
openai
